Create a function called getComputerChoice
This function will serve to retrieve the computers
choice when playing rock, paper, scissors.

Believe it or not, I am super proud of this one.

instead of watching walkthroughs and tutorials and
getting lost, I just used the tools I had and built a
working version myself. I also understand this version.
and could explain it to anyone.

So I am pretty pumped on it.